Generalise locateUnmutedTest into locateTest
